+++ .github/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Improvements
 
-* In Lightning Tensor. Allow `qml.MPSPrep` to accept an MPS with len(MPS) = n_wires-1.
+* In Lightning-Tensor, allow `qml.MPSPrep` to accept an MPS with `len(MPS) = n_wires-1`.
   [(#1064)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1064)
 
 * Lightning devices support dynamically allocated wires (e.g. `qml.device("lightning.qubit")`)